Module 1 Exercises Done

+++ src/module1/_5casting/E03TestOverflow.java
@@ -4,7 +4,7 @@ public class E03TestOverflow {
 	//Exercise 3: These numbers are outside the scope of an int
 	//Modify the code to fix the errors
 	public static void main(String[] args) {
-		int id = 2147483650; // overflow error!
-		int negative = -2147483650; // overflow 
+		long id = 2147483650L; // overflow error!
+		long negative = -2147483650L; // overflow 
 	}
 }
